LD4-98 Refactoring

+++ ledoc-data/src/main/java/dk/ledocsystem/data/model/dashboard/SuperAdminStatistic.java
@@ -11,5 +11,5 @@ public class SuperAdminStatistic {
 
     private UserStat userStat;
 
-    private Long usersOnline;
+    private Integer usersOnline;
 }
